backend/requirements.txt
# Python依赖包
python-docx==1.0.1
requests==2.32.3
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1

# 文档处理依赖
python-pptx==0.6.23
openpyxl==3.1.2
PyPDF2==3.0.1
